Two questions were open. Both are now closed by measurement rather than
argument.

**The heart-rate trend indicator is not worth building.** An interval ride at a
real maximum, lapped at every effort, showed six 20-second sprints out of six
peaking 15-28 s *after* the sprint ended. Simulated second by second, a 20 s
mark reads falling 42% of the time during a sprint -- accurate, since the heart
rate genuinely is decaying from the previous rep, and unreadable as effort
feedback. It would have been correct everywhere and useful nowhere anyone asked
for. The design is kept, unbuilt, with what would change the answer.

**The recovery measurement is a protocol, not a by-product.** The claim that a
pause costs nothing because the end of a ride is already a pause did not
survive: two ordinary rides ceased at 67% and 57% of maximum with no mid-ride
pause at all, and yielded nothing. Docs/RECOVERY-AS-A-MODE.md prices what one
measurement asks against what it returns.

Shipped from it: the lap split, which passes the test the trend mark failed --
it tells a rider something they do not know at a moment they cannot count.

Also here: EffortKit as the shared engine with TrainKit retired, the Service
passing the kernel's 0-3 confidence rather than a boolean, Tools/hr_trend.py,
Ride A and the interval ride preserved as pulled data rather than living only
on the watch, and Docs/INSTALLING.md.

+++ TextKit/src/lib.rs
@@ -429,7 +429,7 @@ mod tests {
             let mut c = Canvas::rect(&mut buf, W, H);
             let m = SEMIBOLD_20_ASCII.draw(&mut c, text, x, 120, align, WHITE);
             let first_lit_col = (0..240).find(|&x| (0..240).any(|y| buf[(y * 240 + x) as usize] != 0)).unwrap();
-            assert_eq!(first_lit_col as i32, expected_left_pen + m.ink_left, "{align:?}");
+            assert_eq!(first_lit_col, expected_left_pen + m.ink_left, "{align:?}");
         }
     }
 
@@ -517,7 +517,7 @@ mod tests {
         let mut buf = vec![0u8; (W * H) as usize];
         let mut c = Canvas::rect(&mut buf, W, H);
         SEMIBOLD_20_ASCII.draw_top(&mut c, "H", 20, 50, Align::Left, WHITE);
-        let first_row = (0..240).find(|&y| (0..240).any(|x| buf[(y * 240 + x) as usize] != 0)).unwrap();
+        let first_row = (0..240).find(|&y| (0..240).any(|x| buf[y * 240 + x] != 0)).unwrap();
         assert!(first_row >= 50 && first_row < 50 + SEMIBOLD_20_ASCII.ascent as usize);
     }
 }
